+++ schemaentry.go
@@ -1,10 +1,18 @@
 package schemaentry
 
 import (
+	log "github.com/Golang-Tools/loggerhelper/v2"
 	jsoniter "github.com/json-iterator/go"
 )
 
 var json = jsoniter.ConfigCompatibleWithStandardLibrary
+var logger *log.Log
+
+func init() {
+	log.Set(log.WithExtFields(log.Dict{"module": "schema-entry-go"}))
+	logger = log.Export()
+	log.Set(log.WithExtFields(log.Dict{}))
+}
 
 //EndPointConfigInterface 叶子节点配置接口
 type EndPointConfigInterface interface {