+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# v2.0.1
+
+## 实现改进
+
++ 改用`github.com/Golang-Tools/loggerhelper/v2`替代`github.com/Golang-Tools/loggerhelper`
++ 不再使用全局log而是使用模块自己的log对象
+
+## bug修复
+
++ 修复`WithConfig`选项设置`WatchMode`无效的问题
+
 # v2.0.0
 
 该版本进行大规模接口变动,使用泛型重构本项目.